samples/basic/weather-agent/python/requirements.txt
python-dotenv
aiohttp
openai
openai-agents
pydantic
azure-identity
microsoft-agents-hosting-core
microsoft-agents-hosting-aiohttp
microsoft-agents-authentication-msal
microsoft-agents-activity
